refactor get-email-from-message
